Add the ViewPager with tabs scenario into the sample app
