+++ sample/kotlin/src/main/AndroidManifest.xml
@@ -36,7 +36,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ViewPagerActivity" />
         <activity android:name=".compose.JetpackComposeActivity" />
 
         <service
