+++ sample/kotlin/src/main/kotlin/com/datadog/android/sample/ViewPagerFragment.kt
@@ -1,0 +1,67 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.sample
+
+import android.app.Activity
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.lifecycle.Lifecycle
+import androidx.viewpager.widget.ViewPager
+import androidx.viewpager2.adapter.FragmentStateAdapter
+import androidx.viewpager2.widget.ViewPager2
+import com.datadog.android.sample.viewpager.FragmentA
+import com.datadog.android.sample.viewpager.FragmentB
+import com.datadog.android.sample.viewpager.FragmentC
+import com.google.android.material.tabs.TabLayout
+import com.google.android.material.tabs.TabLayoutMediator
+
+/**
+ * An [Activity] with a [ViewPager] holding multiple fragments.
+ */
+class ViewPagerFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val root = inflater.inflate(R.layout.fragment_view_pager, container, false)
+        val viewPager: ViewPager2 = root.findViewById(R.id.view_pager)
+        val tabLayout: TabLayout = root.findViewById(R.id.tab_layout)
+        viewPager.adapter = ViewPagerAdapter(parentFragmentManager, lifecycle)
+        TabLayoutMediator(tabLayout, viewPager) { tab, position ->
+            tab.text = "Fragment ${(position + 1)}"
+        }.attach()
+        return root
+    }
+
+    internal inner class ViewPagerAdapter(fragmentManager: FragmentManager, lifecycle: Lifecycle) :
+        FragmentStateAdapter(fragmentManager, lifecycle) {
+
+        override fun getItemCount(): Int {
+            return 3
+        }
+
+        override fun createFragment(position: Int): Fragment {
+            return when (position) {
+                0 -> FragmentA()
+                1 -> FragmentB()
+                else -> FragmentC()
+            }.apply {
+                val args = Bundle().apply {
+                    putString("fragmentClassName", this::class.java.simpleName)
+                    putInt("adapterPosition", position)
+                }
+                arguments = args
+            }
+        }
+    }
+}
