+++ sample/kotlin/src/main/kotlin/com/datadog/android/sample/home/HomeFragment.kt
@@ -60,7 +60,7 @@ internal class HomeFragment :
             R.id.navigation_vitals -> R.id.fragment_vitals
             R.id.navigation_webview -> R.id.fragment_webview
             R.id.navigation_data_list -> R.id.fragment_data_list
-            R.id.navigation_view_pager -> R.id.activity_view_pager
+            R.id.navigation_view_pager -> R.id.fragment_view_pager
             R.id.navigation_picture -> R.id.fragment_picture
             R.id.navigation_compose -> R.id.activity_jetpack_compose
             R.id.navigation_session_replay -> R.id.fragment_session_replay
