+++ sample/kotlin/src/main/res/navigation/nav_graph.xml
@@ -14,8 +14,6 @@
         android:id="@+id/fragment_home"
         android:name="com.datadog.android.sample.home.HomeFragment"
         android:label="@string/title_home">
-        <action android:id="@+id/action_fragment_home_to_activity_view_pager"
-                app:destination="@id/activity_view_pager"/>
     </fragment>
 
     <fragment
@@ -94,8 +92,8 @@
               android:name="com.datadog.android.sample.gdpr.GdprDialogFragment"
               android:label="@string/title_gdpr"/>
 
-    <activity android:id="@+id/activity_view_pager"
-              android:name="com.datadog.android.sample.ViewPagerActivity"
+    <fragment android:id="@+id/fragment_view_pager"
+              android:name="com.datadog.android.sample.ViewPagerFragment"
               android:label="@string/title_view_pager"/>
 
 </navigation>
